pyproject.toml
[tool.poetry]
name = "github-pr-watcher"
version = "1.9.0"
repository = "https://github.com/gm2211/github-watcher"
description = "A MacOs UI tool to monitor GitHub Pull Requests"
authors = ["gm2211 <gmecocci@outlook.com>"]
readme = "README.md"
packages = [{include = "src"}]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
PyQt6 = "6.6.1"
PyQt6-Qt6 = "6.6.1"
PyQt6-sip = "13.6.0"
requests = "^2.31.0"
PyYAML = "^6.0.1"
xxhash = "^3.4.1"

[tool.poetry.group.dev.dependencies]
pillow = "^10.0.0"
pyinstaller = "^6.0.0"

[tool.poetry.scripts]
github-pr-watcher = "github_pr_watcher.main:main"
gpw = "github_pr_watcher.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"